ath10k:  Recover from firmware stuck in scanning state.
CT firmware reports a reson code if we cannot start a
scan, and one of the possible returns is 13, which means
BUSY.  Try to work around stuck scanning state machine by
telling firmware to stop scanning if we detect this out-of-sync
state.

This problem was seen when trying to scan in an area with a very
large amount of APs (400 or so).  With further debugging, hopefully
we can find root cause of the stuck state and fix that too.

Signed-off-by: Ben Greear <greearb@candelatech.com>