ath10k:  fwcfg supports board-file-name.
This is needed because some NICs, like WLE1216V5-20 and WLE1216V5-20
have same PCI and BMI ID, but need different board data.  There is no
programatic way I can think of to differentiate the two reliably in the driver,
so it will be up to users to configure the fwcfg file properly,
perhaps based on MAC address of the radio.

Signed-off-by: Ben Greear <greearb@candelatech.com>